# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TerraCare API",
    "version": "1.0.0",
    "description": "Terariumų, augintinių ir jų priežiūros valdymo API."
  },
  "servers": [
    {
      "url": "/",
      "description": "Dabartinis serveris"
    }
  ],
  "paths": {
...
    "/terrariums/{terrariumId}/pets/{petId}/care-records": {
      "get": {
        "tags": [
          "Priežiūros įrašai"
        ],
        "summary": "Gauti terariumo augintinio priežiūros įrašus",
        "description": "Tikrinama, ar terariumas egzistuoja ir ar augintinis priklauso būtent jam. Grąžinama augintinio priežiūros istorija nuo naujausių įrašų. Jei istorija tuščia, grąžinamas tuščias masyvas.",
        "parameters": [
          {
            "name": "terrariumId",
            "in": "path",
            "required": true,
            "description": "Terariumo ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "petId",
            "in": "path",
            "required": true,
            "description": "Šiam terariumui priklausančio augintinio ID.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Augintinio priežiūros įrašų sąrašas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CareRecord"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas terariumo arba augintinio ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumo ir augintinio ID turi būti teigiami sveikieji skaičiai."
                }
              }
            }
          },
          "404": {
            "description": "Terariumas nerastas arba augintinis šiame terariume nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "terrariumNotFound": {
                    "summary": "Terariumas neegzistuoja",
                    "value": {
                      "error": "Terariumas nerastas."
                    }
                  },
                  "petNotInTerrarium": {
                    "summary": "Augintinis neegzistuoja arba priklauso kitam terariumui",
                    "value": {
                      "error": "Augintinis šiame terariume nerastas."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti augintinio priežiūros įrašų."
                }
              }
            }
          }
        }
      }
    },
    "/terrariums/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Terariumo ID",
          "schema": {
            "type": "integer",
            "minimum": 1,
            "maximum": 2147483647
          }
        }
      ],
      "get": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Gauti terariumą pagal ID",
        "responses": {
          "200": {
            "description": "Terariumo informacija",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terrarium"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas terariumo ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Terariumas nerastas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumas nerastas."
                }
              }
            }
          },
          "500": {
            "$ref": "#/paths/~1terrariums/get/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Atnaujinti terariumą",
        "description": "Pakeičia terariumo informaciją. Reikia pateikti pavadinimą ir visas temperatūros bei drėgmės ribas. Nepateikus last_cleaned ar notes, jie nustatomi į null.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerrariumInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terariumas atnaujintas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terrarium"
                }
              }
            }
          },
          "400": {
            "description": "Netinkamas ID arba netaisyklingas JSON",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/404"
          },
          "422": {
            "$ref": "#/paths/~1terrariums/post/responses/422"
          },
          "500": {
            "$ref": "#/paths/~1terrariums/get/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Ištrinti terariumą",
        "description": "Galima ištrinti tik terariumą, kuriame nėra augintinių.",
        "responses": {
          "204": {
            "description": "Terariumas ištrintas. Atsakymo turinio nėra."
          },
          "400": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/400"
          },
          "404": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/404"
          },
          "409": {
            "description": "Terariume yra augintinių",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Terariumo ištrinti negalima, nes jame yra augintinių."
                }
              }
            }
          },
          "500": {
            "$ref": "#/paths/~1terrariums/get/responses/500"
          }
        }
      }
    },
    "/terrariums/{id}/pets": {
      "get": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Gauti konkretaus terariumo augintinius",
        "description": "Hierarchinis metodas. Jei terariumas egzistuoja, bet neturi augintinių, grąžinamas tuščias masyvas. Jei terariumo nėra, grąžinamas 404.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Terariumo ID",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pasirinkto terariumo augintinių sąrašas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/400"
          },
          "404": {
            "$ref": "#/paths/~1terrariums~1{id}/get/responses/404"
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti terariumo augintinių."
                }
              }
            }
          }
        }
      }
    },
    "/terrariums": {
      "get": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Gauti terariumų sąrašą",
        "description": "Grąžina vieną sąrašo puslapį. Galima ieškoti pagal pavadinimo dalį, neatsižvelgiant į raidžių dydį. Nesant rezultatų, data yra tuščias masyvas.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Puslapio numeris nuo 1. Pernelyg didelis page ir limit derinys atmetamas su 400.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Įrašų skaičius puslapyje.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 5
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Terariumo pavadinimo dalis. Negali būti vien tarpai arba turėti nulinį simbolį.",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "example": "ekspozicija"
          }
        ],
        "responses": {
          "200": {
            "description": "Terariumų sąrašo puslapis su navigacijos nuorodomis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "pagination",
                    "_links"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "allOf": [
                          {
                            "$ref": "#/components/schemas/Terrarium"
                          },
                          {
                            "type": "object",
                            "properties": {
                              "_links": {
                                "type": "object",
                                "properties": {
                                  "self": {
                                    "type": "string",
                                    "example": "/terrariums/1"
                                  },
                                  "pets": {
                                    "type": "string",
                                    "example": "/terrariums/1/pets"
                                  }
                                }
                              }
                            }
                          }
                        ]
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "_links": {
                      "$ref": "#/components/schemas/PageLinks"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Netinkami puslapiavimo arba filtravimo parametrai",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "page ir limit turi būti teigiami sveikieji skaičiai."
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Nepavyko gauti terariumų sąrašo."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Terariumai"
        ],
        "summary": "Sukurti terariumą",
        "description": "Temperatūros ir drėgmės minimumai negali viršyti maksimumų. Nepateikus datos ar pastabų, jos nustatomos į null.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerrariumInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Terariumas sukurtas",
            "headers": {
              "Location": {
                "description": "Sukurto terariumo adresas",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Terrarium"
                }
              }
            }
          },
          "400": {
            "description": "Netaisyklingas JSON formatas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Netaisyklingas JSON formatas."
                }
              }
            }
          },
          "422": {
            "description": "Netinkami duomenys: trūksta laukų, netinkami jų tipai, ribos arba data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Serverio klaida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Terrarium": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Varlių terariumas"
          },
          "temperature_min": {
            "type": "string",
            "example": "22.00"
          },
          "temperature_max": {
            "type": "string",
            "example": "26.00"
          },
          "humidity_min": {
            "type": "string",
            "example": "60.00"
          },
          "humidity_max": {
            "type": "string",
            "example": "80.00"
          },
          "last_cleaned": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Demonstracinis terariumas."
          }
        }
      },
      "TerrariumInput": {
        "type": "object",
        "required": [
          "name",
          "temperature_min",
          "temperature_max",
          "humidity_min",
          "humidity_max"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Pavadinimas negali būti sudarytas tik iš tarpų.",
            "example": "Demo terariumas"
          },
          "temperature_min": {
            "type": "number",
            "minimum": -999.99,
            "maximum": 999.99,
            "description": "Minimali temperatūra °C. Negali viršyti maksimalios.",
            "example": 22
          },
          "temperature_max": {
            "type": "number",
            "minimum": -999.99,
            "maximum": 999.99,
            "description": "Maksimali temperatūra °C.",
            "example": 26
          },
          "humidity_min": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Minimali drėgmė procentais. Negali viršyti maksimalios.",
            "example": 60
          },
          "humidity_max": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Maksimali drėgmė procentais.",
            "example": 80
          },
          "last_cleaned": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-03"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "example": "Demonstracinis terariumas."
          }
        }
      },
      "CareRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "pet_id": {
            "type": "integer",
            "example": 1
          },
          "care_type": {
            "type": "string",
            "enum": [
              "feeding",
              "observation",
              "other"
            ]
          },
          "performed_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "description": "Laiką automatiškai nustato serveris."
          },
          "notes": {
            "type": "string",
            "example": "Augintinis aktyvus, pakitimų nepastebėta."
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Nepavyko gauti augintinių sąrašo."
          }
        }
      }
    }
  }
}